+++ yongwoo/progEx/06/[PROG]-6-yongwoo01.cpp
